[refactor] #19 - 패키지 대소문자 변경을 위한 임시 이름
